+++ js/typescript-library/src/conditions.ts
@@ -1,26 +1,8 @@
-import {BasicMeasure, Field, PACKAGE} from "./index"
+import PACKAGE from "./package";
+import {Condition, BasicMeasure, ConditionType, Field} from "./types";
+import Criteria from "./criteria";
 
 type Primitive = string | number | boolean;
-
-export interface Condition {
-  readonly class: string
-  readonly type: ConditionType
-}
-
-export enum ConditionType {
-  EQ = "EQ",
-  NEQ = "NEQ",
-  LT = "LT",
-  LE = "LE",
-  GT = "GT",
-  GE = "GE",
-  IN = "IN",
-  LIKE = "LIKE",
-  AND = "AND",
-  OR = "OR",
-  NULL = "NULL",
-  NOT_NULL = "NOT_NULL",
-}
 
 function toJSON(c: Condition) {
   return {
@@ -89,17 +71,6 @@ class LogicalCondition implements Condition {
       "one": this.one,
       "two": this.two,
     }
-  }
-}
-
-export class Criteria {
-
-  constructor(public field: Field,
-              public fieldOther: Field,
-              private measure: BasicMeasure,
-              private condition: Condition,
-              public conditionType: ConditionType,
-              public children: Criteria[]) {
   }
 }
