+++ js/typescript-library/src/package.ts
@@ -1,0 +1,2 @@
+const PACKAGE = "io.squashql.query.";
+export default PACKAGE;
