+++ js/typescript-library/src/parameters.ts
@@ -1,4 +1,4 @@
-import {PACKAGE} from "./index"
+import PACKAGE from "./package";
 
 export interface Parameter {
   readonly class: string
